Integrate AST edit distance into evaluation pipeline

+++ evaluation/src/main/java/edu/purdue/dualitylab/evaluation/model/RegexTestSuiteSolution.java
@@ -7,7 +7,7 @@ public record RegexTestSuiteSolution(
         long projectId,
         IndeterminateBoolean fullMatch,
         IndeterminateBoolean partialMatch,
-        double astDistance,
+        int astDistance,
         double autoDistance
 ) {
 }
